+++ demisto_sdk/commands/content_graph/parsers/repository.py
@@ -2,6 +2,7 @@ import multiprocessing
 from pathlib import Path
 from typing import Iterator, List
 
+from demisto_sdk.commands.common.cpu_count import cpu_count
 from demisto_sdk.commands.content_graph.common import PACKS_FOLDER
 from demisto_sdk.commands.content_graph.parsers.pack import PackParser
 
@@ -22,7 +23,7 @@ class RepositoryParser:
             path (Path): The repository path.
         """
         self.path: Path = path
-        pool = multiprocessing.Pool()
+        pool = multiprocessing.Pool(processes=cpu_count())
         self.packs: List[PackParser] = list(pool.map(PackParser, self.iter_packs()))
 
     def iter_packs(self) -> Iterator[Path]: