+++ demisto_sdk/commands/prepare_content/tests/generic_module_unifier_test.py
@@ -3,7 +3,7 @@ import os
 import pytest
 
 from demisto_sdk.commands.common.handlers import JSON_Handler
-from demisto_sdk.commands.unify.generic_module_unifier import GenericModuleUnifier
+from demisto_sdk.commands.prepare_content.generic_module_unifier import GenericModuleUnifier
 from demisto_sdk.tests.test_files.validate_integration_test_valid_types import (DASHBOARD, GENERIC_MODULE,
                                                                                 UNIFIED_GENERIC_MODULE)
 from TestSuite.test_tools import ChangeCWD